+++ docs/.vuepress/config.js
@@ -41,6 +41,10 @@ module.exports = {
       {
         "label": "v1.2.0",
         "key": "v1.2.0"
+      },
+      {
+        "label": "v2.0.0",
+        "key": "v2.0.0"
       }      
     ],
     topbar: {
@@ -92,15 +96,52 @@ module.exports = {
               directory: false,
               path: "/ibc/proto-docs.html"
             },
+            {
+              title: "Roadmap",
+              directory: false,
+              path: "/roadmap/roadmap.html"
+            },
+          ]
+        },
+        {
+          title: "IBC Application Modules",
+          children: [
+            {
+              title: "Interchain Accounts",
+              directory: true,
+              path: "/app_modules",
+              children: [
+                {
+                    title: "Overview",
+                    directory: false,
+                    path: "/app_modules/interchain-accounts/overview.html"
+                }, 
+                {
+                    title: "Integration",
+                    directory: false,
+                    path: "/app_modules/interchain-accounts/integration.html"
+                },
+            ]
+            },
           ]
         },
         {
           title: "Migrations",
           children: [
             {
-              title: "v0.43 SDK to IBC-Go v1.0.0",
+              title: "SDK v0.43 to IBC-Go v1",
               directory: false,
-              path: "/migrations/ibc-migration-043.html"
+              path: "/migrations/sdk-to-v1.html"
+            },
+            {
+              title: "IBC-Go v1 to v2",
+              directory: false,
+              path: "/migrations/v1-to-v2.html"
+            },
+            {
+              title: "IBC-Go v2 to v3",
+              directory: false,
+              path: "/migrations/v2-to-v3.html"
             },
           ]
         },