+++ modules/apps/27-interchain-accounts/host/client/cli/query.go
@@ -1,0 +1,41 @@
+package cli
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/version"
+	"github.com/cosmos/ibc-go/v3/modules/apps/27-interchain-accounts/host/types"
+)
+
+// GetCmdParams returns the command handler for the host submodule parameter querying.
+func GetCmdParams() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "params",
+		Short:   "Query the current interchain-accounts host submodule parameters",
+		Long:    "Query the current interchain-accounts host submodule parameters",
+		Args:    cobra.NoArgs,
+		Example: fmt.Sprintf("%s query interchain-accounts host params", version.AppName),
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.Params(cmd.Context(), &types.QueryParamsRequest{})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res.Params)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}