+++ modules/apps/transfer/keeper/msg_server.go
@@ -7,7 +7,7 @@ import (
 	host "github.com/cosmos/ibc-go/v2/modules/core/24-host"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/ibc-go/v2/modules/apps/transfer/types"
+	"github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
 )
 
 var _ types.MsgServer = Keeper{}